abstracted reusable navigation guard
